feat: Get wallet
Implement get user wallet detail API

(#33350654)[https://github.com/amiraliio/avn-wallet/projects/1#card-33350654]

+++ domain/model/user.go
@@ -6,7 +6,7 @@ import (
 
 type User struct {
 	ID           uuid.UUID      `json:"id" msgpack:"id" validate:"required"`
-	Cellphone    uint           `json:"cellphone" msgpack:"cellphone" validate:"required"`
+	Cellphone    uint64         `json:"cellphone" msgpack:"cellphone" validate:"required"`
 	FirstName    string         `json:"firstName" msgpack:"firstName" validate:"required,gte=3,lte=70"`
 	LastName     string         `json:"lastName" msgpack:"lastName" validate:"required,gte=3,lte=70"`
 	Email        string         `json:"email" msgpack:"email" validate:"required,email"`
